requirements.txt
flask
gunicorn
opencv-contrib-python-headless
pillow
numpy
keras
tensorflow-cpu